chore(refactor): split the big file
